Fixed about a million build errors on Windows.  Someone really needs to get Brian a Windows box.  :)
Fixed a typo bug in loader/limits.c

Added braces to many single-line if-statement bodies.

Fixed a bug in the tilesort SPU related to the scissor rectangle.  This might
fix some of the bugs people reported right before SIGGRAPH.

+++ spu/state_tracker/state_evaluators.c
@@ -98,7 +98,7 @@ void STATE_APIENTRY crStateMap1d (GLenum target, GLdouble u1, GLdouble u2,
 		return;
 	}
 
-	if (order < 1 || order > g->limits.maxEvalOrder)
+	if (order < 1 || order > (int) g->limits.maxEvalOrder)
 	{
 		crStateError(__LINE__, __FILE__, GL_INVALID_VALUE, "Map1d: order oob: %d", order);
 		return;
@@ -161,7 +161,7 @@ void STATE_APIENTRY crStateMap1f (GLenum target, GLfloat u1, GLfloat u2,
 		return;
 	}
 
-	if (order < 1 || order > g->limits.maxEvalOrder)
+	if (order < 1 || order > (int) g->limits.maxEvalOrder)
 	{
 		crStateError(__LINE__, __FILE__, GL_INVALID_VALUE, "Map1f: order oob: %d", order);
 		return;
@@ -230,13 +230,13 @@ void STATE_APIENTRY crStateMap2f(GLenum target, GLfloat u1, GLfloat u2, GLint us
 		return;
 	}
 
-	if (uorder < 1 || uorder > g->limits.maxEvalOrder)
+	if (uorder < 1 || uorder > (int) g->limits.maxEvalOrder)
 	{
 		crStateError(__LINE__, __FILE__, GL_INVALID_VALUE, "Map2f: order oob: %d", uorder);
 		return;
 	}
 
-	if (vorder < 1 || vorder > g->limits.maxEvalOrder)
+	if (vorder < 1 || vorder > (int) g->limits.maxEvalOrder)
 	{
 		crStateError(__LINE__, __FILE__, GL_INVALID_VALUE, "Map2f: vorder oob: %d", vorder);
 		return;
@@ -318,13 +318,13 @@ void STATE_APIENTRY crStateMap2d(GLenum target, GLdouble u1, GLdouble u2, GLint 
 		return;
 	}
 
-	if (uorder < 1 || uorder > g->limits.maxEvalOrder)
+	if (uorder < 1 || uorder > (int) g->limits.maxEvalOrder)
 	{
 		crStateError(__LINE__, __FILE__, GL_INVALID_VALUE, "Map2d: order oob: %d", uorder);
 		return;
 	}
 
-	if (vorder < 1 || vorder > g->limits.maxEvalOrder)
+	if (vorder < 1 || vorder > (int) g->limits.maxEvalOrder)
 	{
 		crStateError(__LINE__, __FILE__, GL_INVALID_VALUE, "Map2d: vorder oob: %d", vorder);
 		return;
