+++ spu/state_tracker/state_texture.c
@@ -894,7 +894,7 @@ void STATE_APIENTRY crStateTexImage1D (GLenum target, GLint level, GLint compone
 		return;
 	}
 
-	if (width > g->limits.maxTextureSize)
+	if (width > (int) g->limits.maxTextureSize)
 	{
 		if (target == GL_PROXY_TEXTURE_1D)
 		{
@@ -1098,13 +1098,13 @@ void STATE_APIENTRY crStateTexImage2D (GLenum target, GLint level, GLint compone
 
 	if (target == GL_TEXTURE_2D)
 	{
-		if (width > g->limits.maxTextureSize)
+		if (width > (int) g->limits.maxTextureSize)
 		{
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 				     "glTexImage2D width oob: %d", width);
 			return;
 		}
-		if (height > g->limits.maxTextureSize)
+		if (height > (int) g->limits.maxTextureSize)
 		{
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 				     "glTexImage2D height oob: %d", height);
@@ -1113,7 +1113,7 @@ void STATE_APIENTRY crStateTexImage2D (GLenum target, GLint level, GLint compone
 	}
 	else if (target == GL_PROXY_TEXTURE_2D)
 	{
-		if (width > g->limits.maxTextureSize || height > g->limits.maxTextureSize)
+		if (width > (int) g->limits.maxTextureSize || height > (int) g->limits.maxTextureSize)
 		{
 			/* clear all the texture object state */
 			crStateTextureInitTextureObj(t, &(t->proxy2D), 0, GL_TEXTURE_2D);
@@ -1123,8 +1123,8 @@ void STATE_APIENTRY crStateTexImage2D (GLenum target, GLint level, GLint compone
 #ifdef CR_ARB_texture_cube_map
 	else if (target == GL_PROXY_TEXTURE_CUBE_MAP_ARB)
 	{
-		if (width > g->limits.maxCubeMapTextureSize ||
-			height > g->limits.maxCubeMapTextureSize ||
+		if (width > (int) g->limits.maxCubeMapTextureSize ||
+			height > (int) g->limits.maxCubeMapTextureSize ||
 			width != height)
 		{
 			/* clear all the texture object state */
@@ -1135,13 +1135,13 @@ void STATE_APIENTRY crStateTexImage2D (GLenum target, GLint level, GLint compone
 	}
 	else // Cube map
 	{
-		if (width > g->limits.maxCubeMapTextureSize)
+		if (width > (int) g->limits.maxCubeMapTextureSize)
 		{
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 				     "glTexImage2D width oob: %d", width);
 			return;
 		}
-		if (height > g->limits.maxCubeMapTextureSize)
+		if (height > (int) g->limits.maxCubeMapTextureSize)
 		{
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 				     "glTexImage2D height oob: %d", height);
@@ -1286,7 +1286,7 @@ void STATE_APIENTRY crStateTexSubImage1D (GLenum target, GLint level, GLint xoff
 		return;
 	}
 
-	if (width > g->limits.maxTextureSize)
+	if (width > (int) g->limits.maxTextureSize)
 	{
 		crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 					"glSubTexImage1D width oob: %d", width);
@@ -1361,13 +1361,13 @@ void STATE_APIENTRY crStateTexSubImage2D (GLenum target, GLint level, GLint xoff
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE, "glTexSubImage2D level oob: %d", level);
 			return;
 		}
-		if (width < 0 || width > g->limits.maxTextureSize)
+		if (width < 0 || width > (int) g->limits.maxTextureSize)
 		{
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 				     "glSubTexImage2D width oob: %d", width);
 			return;
 		}
-		if (height < 0 || height > g->limits.maxTextureSize)
+		if (height < 0 || height > (int) g->limits.maxTextureSize)
 		{
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 				     "glSubTexImage2D height oob: %d", height);
@@ -1382,13 +1382,13 @@ void STATE_APIENTRY crStateTexSubImage2D (GLenum target, GLint level, GLint xoff
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE, "glTexSubImage2D level oob: %d", level);
 			return;
 		}
-		if (width < 0 || width > g->limits.maxCubeMapTextureSize)
+		if (width < 0 || width > (int) g->limits.maxCubeMapTextureSize)
 		{
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 				     "glTexSubImage2D width oob: %d", width);
 			return;
 		}
-		if (height < 0 || height > g->limits.maxCubeMapTextureSize)
+		if (height < 0 || height > (int) g->limits.maxCubeMapTextureSize)
 		{
 			crStateError(__LINE__, __FILE__, GL_INVALID_VALUE,
 				     "glSubTexImage2D height oob: %d", height);
